pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "factorminer"
version = "0.1.0"
description = "LLM-powered quantitative factor mining with evolutionary search"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "FactorMiner Team"},
]
keywords = ["quantitative-finance", "factor-mining", "llm", "alpha-research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "bottleneck>=1.4",
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "torch>=2.0",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "tqdm>=4.65",
    "statsmodels>=0.14.0",
    "sympy>=1.13",
    "networkx>=3.0",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x; platform_system == 'Linux'",
]
llm = [
    "openai",
    "anthropic",
    "google-generativeai",
    "sentence-transformers>=2.0",
    "faiss-cpu>=1.7",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]
all = [
    "cupy-cuda12x; platform_system == 'Linux'",
    "openai",
    "anthropic",
    "google-generativeai",
    "sentence-transformers>=2.0",
    "faiss-cpu>=1.7",
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
factorminer = "factorminer.cli:main"

[project.urls]
Homepage = "https://github.com/minihellboy/factorminer"
Repository = "https://github.com/minihellboy/factorminer"
Issues = "https://github.com/minihellboy/factorminer/issues"

[tool.setuptools.packages.find]
include = ["factorminer*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = [
    # Quant research code uses matrix notation such as M, T, X, and K.
    "N803",
    "N806",
    # Formula catalogs and LLM prompts are intentionally long string literals.
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"run_demo.py" = ["E402"]

[tool.pytest.ini_options]
testpaths = ["factorminer/tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
